move rendering utils to its own folder
